Frankie and Carl

+++ harry-potter/09102010-blundell-xavi/src/main/java/silent/BookStore.java
@@ -6,11 +6,11 @@ public class BookStore {
         return new Order();
     }
 
-    public int purchase(Order order) {
+    public double purchase(Order order) {
         return 8 * order.getTotalBooks() + applyDiscount(order);
     }
 
-    private int applyDiscount(Order order) {
+    private double applyDiscount(Order order) {
         DiscountCalculator discountCalculator = new DiscountCalculator(order);
         return discountCalculator.calculate();
     }
